feat: Add tracing for Logger; A little ref for colors
