--- src/core/logger/constants/default-colors.const.ts
@@ -1,7 +1,0 @@
-export const DefaultColors = {
-  white: "#ffffff",
-  light_red: "#ff9999",
-  light_green: "#adff99",
-  light_cyan: "#99d3ff",
-  light_yellow: "#fcff99",
-} as const;
